refactor(agent): move sandboxed bash tool to dedicated file
